requirements.txt
streamlit
google-genai
pypdf
python-dotenv
pandas
